Modify from App engine to Compute app
App engine did not give enough control to prevent memory
related issues, so we migrated onto compute engine to have more control
at the cost of managing more of the infrastructure ourselves. In reality
this was probably for the best since we have very specific requirements.

Gunicorn.py includes some default configuration from the official docs,
and also loads the model on server start to not overwhelm the servers
with multiple copies